# Add signal handler to reap zombie process.